requirements.txt
torch
cv2-enumerate-cameras



